% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_irish30}
\alias{data_corpus_irish30}
\title{Irish Dail speeches from 2007-2011}
\format{The corpus object for the 30th Dail speeches, with document-level 
  variables for speaker, debate, and speaker's party.}
\source{
Benjamin E. Lauderdale and Alexander Herzog.  2016.
  "\href{https://www.cambridge.org/core/journals/political-analysis/article/measuring-political-positions-from-legislative-speech/35D8B53C4B7367185325C25BBE5F42B4}{Measuring
   Political Positions from Legislative Speech}." \emph{Political Analysis}
  24 (3, July): 374-394.
}
\usage{
data_corpus_irish30
}
\description{
Sampled speeches and document-level variables from debates in the 30th Dail.
Only ten debates are included.
}
\examples{
summary(data_corpus_irish30, 5)
}
\keyword{datasets}
