% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_wordshoal.R
\name{print.textmodel_wordshoal}
\alias{print.textmodel_wordshoal}
\title{Print method for textmodel_wordshoal}
\usage{
\method{print}{textmodel_wordshoal}(x, ...)
}
\arguments{
\item{x}{for print method, the object to be printed}

\item{...}{additional arguments passed to \code{\link{print}}}
}
\description{
Provides a print method for this class of object.
}
\keyword{internal}
