% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_wordshoal.R
\name{summary.textmodel_wordshoal}
\alias{summary.textmodel_wordshoal}
\title{Summarize a fitted textmodel_wordshoal object.}
\usage{
\method{summary}{textmodel_wordshoal}(object, ...)
}
\arguments{
\item{object}{results of \code{\link{textmodel_wordshoal}} to be summarized}

\item{...}{additional arguments passed to \code{print}}
}
\description{
\code{summary} method for a fitted \code{\link{textmodel_wordshoal}} object.
}
