% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_wordshoal.R
\name{summary.textmodel_wordshoal_fitted}
\alias{summary.textmodel_wordshoal_fitted}
\title{Summarize a fitted textmodel_wordshoal object.}
\usage{
\method{summary}{textmodel_wordshoal_fitted}(object, ...)
}
\description{
\code{summary} method for a fitted \code{\link{textmodel_wordshoal}} object.
}
