% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{wordshoal-package}
\alias{wordshoal}
\alias{wordshoal-package}
\title{The Lauderdale and Herzog (2016) "wordshoal" text scaling model}
\description{
An extension to the \pkg{quanteda} providing a `textmodel` for the 
"wordshoal" method describe in Lauderdale and Herzog (2016).
}
\references{
Benjamin E. Lauderdale and Alexander Herzog.  2016.
  "\href{https://www.cambridge.org/core/journals/political-analysis/article/measuring-political-positions-from-legislative-speech/35D8B53C4B7367185325C25BBE5F42B4}{Measuring
   Political Positions from Legislative Speech}." \emph{Political Analysis}
  24 (3, July): 374-394.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/kbenoit/wordshoal}
  \item Report bugs at \url{https://github.com/kbenoit/wordshoal/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} [contributor]

Authors:
\itemize{
  \item Benjamin Lauderdale \email{B.E.lauderdale@lse.ac.uk}
}

Other contributors:
\itemize{
  \item Akitaka Matsuo \email{a.matsuo@lse.ac.uk} [contributor]
}

}
